Fix checkpoint.sh to check for strict mode warnings in dev docs build
The script was only checking for warnings in user docs (mkdocs.yml)
but not in developer docs (mkdocs-dev.yml). This allowed warnings
to slip through that would fail on GitHub deployment.

Now both builds check for strict mode warnings:
- contains an unrecognized relative link
- does not contain an anchor
- contains an absolute link

This fix needs to be on ALL branches to prevent merge conflicts.